requirements.txt
# Discord Bot
discord.py==2.4.0
python-dotenv==1.0.0

# Web Server (for Render hosting)
Flask==3.0.0
Werkzeug==3.0.1

# API & Data
requests==2.32.3

# Scheduler
APScheduler==3.10.4

# Database
# sqlite3 is included in Python standard library

# Math & Utilities
numpy==1.26.0
scikit-learn==1.3.2

# NLP & ML (optional - for semantic search)
sentence-transformers==2.2.2
